Finished finxing semaphores May 17
Fixed erronous behavior in recordFrames. To check if the next frame had been captured, the wrong values were being compared. The logic to check for semaphore signaling was also cleaned up to be more readable and work properly.
The method to get the timestamp of the last frame was also changed to be `pxd_capturedSysTicks` instead of `pxd_buffersSysTicks`, which seems to be more accurate.