Finished adding context cameras
An large amount of testing was done for what valid frame rates and resolutions will not max out the CPU during a capture. If nothing is open other than the application, the two context cameras can run at 640x480 and 15 fps or 32x240 at 30 fps. Anything higher causes large numbers of frames to be dropped. The exposure time of the IR camera has also been updated so that the framerates will match. IMU integration testing will commence after this commit.