Added recursive directory creation
Functionality was added so that when a directory is created, parent directories that do not exist will be created as well. This will prevent files from being unable to save, though it may be in an unexpected place if the path was specified incorrectly.
Initital framework for refactoring record thread written.