Number updates
Several variables were updated to fix issues encountered. The maxPortNumber was increased for the shutter, which allows it to be controlled, but some hangs occur. The path for single images was updated. The starting value in the calibration run was also updated so that only exposure times within the cameras operating range was used.